back/requirements.txt
asgiref==3.8.1
certifi==2025.7.9
charset-normalizer==3.4.2
dj-database-url==2.3.0
Django==5.0.7
django-cleanup==9.0.0
django-cors-headers==4.4.0
django-filter==24.2
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
Markdown==3.6
pillow==10.4.0
psycopg2-binary==2.9.10
pydot==3.0.2
PyJWT==2.9.0
pyparsing==3.2.0
python-decouple==3.8
requests==2.32.4
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
whitenoise==6.8.2